backend/requirements.txt
# --- Phase 1: Core ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20

# --- Phase 1: Database ---
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# --- Phase 2: Auth ---
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# --- Phase 3: PDF Pipeline ---
celery[redis]==5.4.0
bcrypt==4.2.1
PyMuPDF==1.25.3
langchain-text-splitters==0.3.4
google-genai==1.14.0
qdrant-client==1.12.1
psycopg2-binary==2.9.10

# --- Phase 4: Query ---
cohere==5.13.3
